requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8
django-filter==23.5
drf-spectacular==0.27.1
channels==4.0.0
channels-redis==4.1.0
celery==5.3.6
redis==5.0.1
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
